more hg-engine setup

+++ Main/Forms/HgeIvEvForm.cs
@@ -2,10 +2,10 @@
 {
     public partial class HgeIvEvForm : Form
     {
-        public ushort[] IVs = new ushort[6];
-        public ushort[] EVs = new ushort[6];
+        public byte[] IVs = new byte[6];
+        public byte[] EVs = new byte[6];
 
-        public HgeIvEvForm(ushort[] ivs, ushort[] evs)
+        public HgeIvEvForm(byte[] ivs, byte[] evs)
         {
             this.IVs = ivs;
             this.EVs = evs;
@@ -36,22 +36,22 @@
 
         private void SaveIvValues()
         {
-            IVs[0] = (ushort)num_IvHp.Value;
-            IVs[1] = (ushort)num_IvAtk.Value;
-            IVs[2] = (ushort)num_IvDef.Value;
-            IVs[3] = (ushort)num_IvSpd.Value;
-            IVs[4] = (ushort)num_IvSpAtk.Value;
-            IVs[5] = (ushort)num_IvSpDef.Value;
+            IVs[0] = (byte)num_IvHp.Value;
+            IVs[1] = (byte)num_IvAtk.Value;
+            IVs[2] = (byte)num_IvDef.Value;
+            IVs[3] = (byte)num_IvSpd.Value;
+            IVs[4] = (byte)num_IvSpAtk.Value;
+            IVs[5] = (byte)num_IvSpDef.Value;
         }
 
         private void SaveEvValues()
         {
-            EVs[0] = (ushort)num_EvHp.Value;
-            EVs[1] = (ushort)num_EvAtk.Value;
-            EVs[2] = (ushort)num_EvDef.Value;
-            EVs[3] = (ushort)num_EvSpd.Value;
-            EVs[4] = (ushort)num_EvSpAtk.Value;
-            EVs[5] = (ushort)num_EvSpDef.Value;
+            EVs[0] = (byte)num_EvHp.Value;
+            EVs[1] = (byte)num_EvAtk.Value;
+            EVs[2] = (byte)num_EvDef.Value;
+            EVs[3] = (byte)num_EvSpd.Value;
+            EVs[4] = (byte)num_EvSpAtk.Value;
+            EVs[5] = (byte)num_EvSpDef.Value;
         }
 
         private void HgeIvEvForm_FormClosing(object sender, FormClosingEventArgs e)
